refactor: moved duplicated code to commons
Refs: XRDDEV-1960
