+++ src/central-server/admin-service/infra-api-rest/src/main/resources/META-INF/custom-constraints.xml
@@ -9,7 +9,7 @@
 
     <bean class="InitialServerConfDto" ignore-annotations="false">
         <field name="instanceIdentifier">
-            <constraint annotation="org.niis.xroad.cs.admin.rest.api.openapi.validator.IdentifierChars">
+            <constraint annotation="org.niis.xroad.restapi.openapi.validator.IdentifierChars">
             </constraint>
         </field>
         <field name="centralServerAddress">
@@ -42,39 +42,39 @@
 
     <bean class="MemberAddDto" ignore-annotations="false">
         <field name="memberName">
-            <constraint annotation="org.niis.xroad.cs.admin.rest.api.openapi.validator.NoControlChars"></constraint>
+            <constraint annotation="org.niis.xroad.restapi.openapi.validator.NoControlChars"></constraint>
         </field>
     </bean>
     <bean class="MemberNameDto" ignore-annotations="false">
         <field name="memberName">
-            <constraint annotation="org.niis.xroad.cs.admin.rest.api.openapi.validator.NoControlChars"></constraint>
+            <constraint annotation="org.niis.xroad.restapi.openapi.validator.NoControlChars"></constraint>
         </field>
     </bean>
 
     <bean class="MemberClassDto" ignore-annotations="false">
         <field name="code">
-            <constraint annotation="org.niis.xroad.cs.admin.rest.api.openapi.validator.IdentifierChars"></constraint>
+            <constraint annotation="org.niis.xroad.restapi.openapi.validator.IdentifierChars"></constraint>
         </field>
     </bean>
 
     <bean class="NewMemberIdDto" ignore-annotations="false">
         <field name="memberClass">
-            <constraint annotation="org.niis.xroad.cs.admin.rest.api.openapi.validator.IdentifierChars"></constraint>
+            <constraint annotation="org.niis.xroad.restapi.openapi.validator.IdentifierChars"></constraint>
         </field>
         <field name="memberCode">
-            <constraint annotation="org.niis.xroad.cs.admin.rest.api.openapi.validator.IdentifierChars"></constraint>
+            <constraint annotation="org.niis.xroad.restapi.openapi.validator.IdentifierChars"></constraint>
         </field>
     </bean>
 
     <bean class="NewSubsystemIdDto" ignore-annotations="false">
         <field name="subsystemCode">
-            <constraint annotation="org.niis.xroad.cs.admin.rest.api.openapi.validator.IdentifierChars"></constraint>
+            <constraint annotation="org.niis.xroad.restapi.openapi.validator.IdentifierChars"></constraint>
         </field>
         <field name="memberClass">
-            <constraint annotation="org.niis.xroad.cs.admin.rest.api.openapi.validator.IdentifierChars"></constraint>
+            <constraint annotation="org.niis.xroad.restapi.openapi.validator.IdentifierChars"></constraint>
         </field>
         <field name="memberCode">
-            <constraint annotation="org.niis.xroad.cs.admin.rest.api.openapi.validator.IdentifierChars"></constraint>
+            <constraint annotation="org.niis.xroad.restapi.openapi.validator.IdentifierChars"></constraint>
         </field>
     </bean>
